rename mmf.* files to mmftargets.*
